Move Redis endpoint validators to utility
