+++ instrumentation-api-incubator/src/main/java/io/opentelemetry/instrumentation/api/incubator/semconv/db/internal/RedisServerTargetUtil.java
@@ -1,0 +1,92 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.api.incubator.semconv.db.internal;
+
+import java.net.URI;
+import java.net.URISyntaxException;
+
+// TODO: Remove this class and switch callers to DbServerEndpointUtil once #20015 merges.
+final class RedisServerTargetUtil {
+
+  static boolean isIpv6Literal(String value) {
+    int zoneStart = value.indexOf('%');
+    String literal = zoneStart < 0 ? value : value.substring(0, zoneStart);
+    if (zoneStart >= 0 && !isZoneId(value.substring(zoneStart + 1))) {
+      return false;
+    }
+    for (int i = 0; i < literal.length(); i++) {
+      char c = literal.charAt(i);
+      if (c != ':' && c != '.' && !isAsciiHexDigit(c)) {
+        return false;
+      }
+    }
+    int ipv4Start = literal.lastIndexOf(':') + 1;
+    if (literal.indexOf('.') >= 0 && !isIpv4Literal(literal.substring(ipv4Start))) {
+      return false;
+    }
+    try {
+      return new URI("db://[" + literal + "]").getHost() != null;
+    } catch (URISyntaxException ignored) {
+      return false;
+    }
+  }
+
+  static boolean isIpv4Literal(String value) {
+    int parts = 0;
+    int digits = 0;
+    int number = 0;
+    for (int i = 0; i <= value.length(); i++) {
+      char c = i == value.length() ? '.' : value.charAt(i);
+      if (c == '.') {
+        if (digits == 0 || (digits > 1 && value.charAt(i - digits) == '0') || number > 255) {
+          return false;
+        }
+        parts++;
+        digits = 0;
+        number = 0;
+      } else {
+        if (!isAsciiDigit(c)) {
+          return false;
+        }
+        if (++digits > 3) {
+          return false;
+        }
+        number = number * 10 + c - '0';
+      }
+    }
+    return parts == 4;
+  }
+
+  private static boolean isZoneId(String value) {
+    if (value.isEmpty()) {
+      return false;
+    }
+    for (int i = 0; i < value.length(); i++) {
+      if (!isUnreserved(value.charAt(i))) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private static boolean isUnreserved(char c) {
+    return c == '-' || c == '.' || c == '_' || c == '~' || isAsciiLetterOrDigit(c);
+  }
+
+  private static boolean isAsciiHexDigit(char c) {
+    return (c >= 'a' && c <= 'f') || (c >= 'A' && c <= 'F') || isAsciiDigit(c);
+  }
+
+  private static boolean isAsciiLetterOrDigit(char c) {
+    return (c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') || isAsciiDigit(c);
+  }
+
+  private static boolean isAsciiDigit(char c) {
+    return c >= '0' && c <= '9';
+  }
+
+  private RedisServerTargetUtil() {}
+}
